vpp/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
markdown-it-py==4.0.0
mdit-py-plugin==0.5.0
linkify-it-py==2.0.
uc-micro-py==1.0.3